# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: WhatsApp SaaS Platform API
  description: |
    REST API for the WhatsApp SaaS Platform - A multi-tenant WhatsApp Business API platform
    for managing conversations, bookings, and AI-powered customer interactions.

    ## Features
    - Multi-tenant salon management
    - WhatsApp message webhooks
    - AI-powered conversation management
    - Booking system integration
    - Real-time analytics and metrics
    - Rate limiting and caching

    ## Authentication
    Admin endpoints require an `x-admin-token` header with your admin API token.

    ## Rate Limiting
    - Webhook endpoints: 100 requests/minute per IP
    - Admin endpoints: 30 requests/minute per IP
    - Rate limits are enforced per endpoint category

    ## Caching
    - Health checks: 30 seconds
    - Bookings: 1 minute
    - Messages: 1 minute
    - Stats: 2 minutes
    - AI Analytics: 15 minutes
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
    url: https://example.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com
    description: Production server
  - url: https://staging-api.example.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development server
...
security:
  - AdminToken: []
paths:
...
  /webhook:
    get:
      summary: Webhook verification endpoint
      description: |
        WhatsApp webhook verification endpoint. Meta Platform calls this endpoint during
        webhook setup to verify ownership. Responds with the challenge parameter if the
        verify token matches.
      tags:
        - Webhooks
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
            enum: [subscribe]
          description: Must be 'subscribe'
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
          description: Verification token configured in Meta Developer Portal
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
          description: Challenge string to echo back
      responses:
        '200':
          description: Webhook verified successfully
          content:
            text/plain:
              schema:
                type: string
                description: The challenge parameter
        '403':
          description: Verification failed - invalid token or parameters

    post:
      summary: Webhook receiver endpoint
      description: |
        Receives WhatsApp webhook events from Meta Platform. Handles incoming messages,
        status updates, and other WhatsApp Business API events.

        **Security:**
        - HMAC signature verification (x-hub-signature-256 header)
        - Rate limited to 100 requests/minute per IP
        - Returns 200 ACK immediately, processes events asynchronously
      tags:
        - Webhooks
      parameters:
        - name: x-hub-signature-256
          in: header
          required: true
          schema:
            type: string
          description: HMAC SHA-256 signature of request body
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEvent'
            example:
              object: whatsapp_business_account
              entry:
                - id: '123456789'
                  changes:
                    - value:
                        messaging_product: whatsapp
                        metadata:
                          display_phone_number: '+1234567890'
                          phone_number_id: '987654321'
                        messages:
                          - from: '1234567890'
                            id: msg_123abc
                            timestamp: '1642512000'
                            type: text
                            text:
                              body: 'I would like to book an appointment tomorrow at 2pm'
      responses:
        '200':
          description: Event received and queued for processing
          content:
            text/plain:
              schema:
                type: string
                example: EVENT_RECEIVED
        '401':
          description: Invalid HMAC signature
...
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: Admin API token for authenticating administrative endpoints
...
  schemas:
...
    WebhookEvent:
      type: object
      required:
        - object
        - entry
      properties:
        object:
          type: string
          example: whatsapp_business_account
        entry:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              changes:
                type: array
                items:
                  type: object
                  properties:
                    value:
                      type: object
                      properties:
                        messaging_product:
                          type: string
                          example: whatsapp
                        metadata:
                          type: object
                          properties:
                            display_phone_number:
                              type: string
                            phone_number_id:
                              type: string
                        messages:
                          type: array
                          items:
                            type: object
                            properties:
                              from:
                                type: string
                              id:
                                type: string
                              timestamp:
                                type: string
                              type:
                                type: string
                                enum: [text, image, video, audio, document, location, contacts]
                              text:
                                type: object
                                properties:
                                  body:
                                    type: string
...